requirements.txt
black==23.12.1
click==8.1.7
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
reindent==3.5.1
RPi.GPIO==0.7.1
